refactor(cms): split page content query
Move the generic page-content D1 reader into cms/page.ts so cms/index.ts keeps public exports, editor snapshot orchestration, and project fallback queries.

Verified with pnpm --filter @anipotts/lib typecheck, pnpm --filter @anipotts/lib test, pnpm --filter @anipotts/lib build, pnpm exec prettier --check packages/lib/src/cms/index.ts packages/lib/src/cms/page.ts, git diff --check -- packages/lib/src/cms/index.ts packages/lib/src/cms/page.ts, and pnpm dlx knip --no-exit-code.
